Reorganized into namespaced package
